incorporated an arm class into auto WARNING: CHANGED TO23, MAY CAUSE UNFORESEEN ERRORS HOPEFULLY NOT
