+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/SampleMecanumDrive.java
@@ -74,7 +74,6 @@ public class SampleMecanumDrive extends MecanumDrive {
 
     private DcMotorEx leftFront, leftRear, rightRear, rightFront;
     private List<DcMotorEx> motors;
-    private Servo cone, upArm, popper, light, lowArm;
 
     private IMU imu;
     private VoltageSensor batteryVoltageSensor;
@@ -100,12 +99,6 @@ public class SampleMecanumDrive extends MecanumDrive {
         leftRear = hardwareMap.get(DcMotorEx.class, "leftRear");
         rightRear = hardwareMap.get(DcMotorEx.class, "rightRear");
         rightFront = hardwareMap.get(DcMotorEx.class, "rightFront");
-
-        lowArm = hardwareMap.servo.get("rotator"); //0e
-        upArm = hardwareMap.servo.get("extender"); //0
-        cone = hardwareMap.servo.get("cone"); //2
-        popper = hardwareMap.servo.get("popper"); //3
-        light = hardwareMap.servo.get("release");//1e
 
         motors = Arrays.asList(leftFront, leftRear, rightRear, rightFront);
 
@@ -141,26 +134,6 @@ public class SampleMecanumDrive extends MecanumDrive {
                 follower, HEADING_PID, batteryVoltageSensor,
                 lastEncPositions, lastEncVels, lastTrackingEncPositions, lastTrackingEncVels
         );
-    }
-    public void servosGoToUpright() {
-        cone.setPosition(0.6);
-        upArm.setPosition(0.64);
-        lowArm.setPosition(0.26 + 0.07);
-        light.setPosition(0);
-    }
-    public void servosGoToPickup() {
-        cone.setPosition(0.53);
-        upArm.setPosition(0.65);
-        lowArm.setPosition(0.01);
-        light.setPosition(0);
-    }
-    public void servosGoToPickup(double c, double u, double l) {
-        cone.setPosition(c);
-        upArm.setPosition(u);
-        lowArm.setPosition(l);
-    }
-    public void operateClaw(int state) {
-        popper.setPosition(0.13 + state * 0.22);
     }
     public TrajectoryBuilder trajectoryBuilder(Pose2d startPose) {
         return new TrajectoryBuilder(startPose, VEL_CONSTRAINT, ACCEL_CONSTRAINT);
